Split abstract.py into a new package
