+++ ogr/abstract/__init__.py
@@ -1,0 +1,37 @@
+# Copyright Contributors to the Packit project.
+# SPDX-License-Identifier: MIT
+
+from ogr.abstract.abstract_class import OgrAbstractClass
+from ogr.abstract.access_level import AccessLevel
+from ogr.abstract.auth_method import AuthMethod
+from ogr.abstract.comment import (
+    AnyComment,
+    Reaction,
+    Comment,
+    IssueComment,
+    CommitComment,
+    PRComment
+)
+
+from ogr.abstract.commit_flag import (
+    CommitFlag
+)
+
+from ogr.abstract.git_project import GitProject
+from ogr.abstract.git_service import GitService
+from ogr.abstract.git_tag import GitTag
+from ogr.abstract.git_user import GitUser
+from ogr.abstract.issue import Issue
+from ogr.abstract.label import (
+    Label,
+    PRLabel,
+    IssueLabel
+)
+from ogr.abstract.pull_request import PullRequest
+from ogr.abstract.release import Release
+from ogr.abstract.status import (
+    IssueStatus,
+    PRStatus,
+    CommitStatus,
+    MergeCommitStatus
+)
